+++ scripts/auth-routes.js
